Move all Kafka library instrumenter parts to internal package
