+++ instrumentation/kafka-clients/kafka-clients-2.6/library/src/main/java/io/opentelemetry/instrumentation/kafkaclients/KafkaTracing.java
@@ -13,8 +13,8 @@ import io.opentelemetry.context.propagation.TextMapGetter;
 import io.opentelemetry.context.propagation.TextMapPropagator;
 import io.opentelemetry.context.propagation.TextMapSetter;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.instrumentation.kafka.KafkaConsumerRecordGetter;
-import io.opentelemetry.instrumentation.kafka.KafkaHeadersSetter;
+import io.opentelemetry.instrumentation.kafka.internal.KafkaConsumerRecordGetter;
+import io.opentelemetry.instrumentation.kafka.internal.KafkaHeadersSetter;
 import java.util.concurrent.Future;
 import java.util.function.BiFunction;
 import org.apache.kafka.clients.consumer.Consumer;
