+++ instrumentation/spring/spring-kafka-2.7/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/kafka/KafkaBatchProcessSpanLinksExtractor.java
@@ -9,8 +9,8 @@ import io.opentelemetry.context.Context;
 import io.opentelemetry.context.propagation.ContextPropagators;
 import io.opentelemetry.instrumentation.api.instrumenter.SpanLinksBuilder;
 import io.opentelemetry.instrumentation.api.instrumenter.SpanLinksExtractor;
-import io.opentelemetry.instrumentation.kafka.KafkaConsumerIteratorWrapper;
-import io.opentelemetry.instrumentation.kafka.KafkaConsumerRecordGetter;
+import io.opentelemetry.instrumentation.kafka.internal.KafkaConsumerIteratorWrapper;
+import io.opentelemetry.instrumentation.kafka.internal.KafkaConsumerRecordGetter;
 import java.util.Iterator;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
